In sync mode, execute tasks in a depth first fashion
Since we execute tasks in a single thread in sync mode, executing tasks
breadth first fills up the task queue rather quickly while running
big crawlers that loop through a lot of data. Instead we can execute tasks
depth first without putting tasks on the queue.

We check whether the execution stack is too deep on every execution so
that we don't get stuck in a recursive loop or go too deep in a single
branch of execution.