Fix failing test as the source has changed (#224)
This test relies on a web page that seems to have changed (the author meta tag isn't part of the page source anymore). While the author is still mentioned on the page, the XPath selector would be more complex and would probably fail again soon anyways. As it's not relevant for this particular test, I've simply removed this property.

A more reliable long-term solution would be to store the source of web pages that the tests rely on as part of this repository and then mock the HTTP requests.